refactor(webapp): consolidate theme into lib/theme.ts
