+++ packages/webapp/src/app/App.tsx
@@ -9,22 +9,17 @@ import { DevToolPanel, isDevToolsEnabled } from '@/components-v2/DevToolPanel';
 import { useMeta } from '@/hooks/useMeta';
 import { useUser } from '@/hooks/useUser';
 import { useStore } from '@/store';
-import { useThemeStore } from '@/store/theme';
 import { globalEnv } from '@/utils/env';
 import { LocalStorageKeys } from '@/utils/local-storage';
-import { applyTheme } from '@/utils/theme';
+import { useTheme } from '@/lib/theme';
 
 import 'react-toastify/dist/ReactToastify.css';
 
 const App = () => {
     const env = useStore((state) => state.env);
     const setShowGettingStarted = useStore((state) => state.setShowGettingStarted);
-    const darkMode = useThemeStore((state) => state.darkMode);
-
-    // Sync theme state to DOM so CSS tokens respond to dark/light mode
-    useEffect(() => {
-        applyTheme(darkMode);
-    }, [darkMode]);
+    // Sync persisted theme preference to the DOM
+    useTheme();
     const [_, setLastEnvironment] = useLocalStorage(LocalStorageKeys.LastEnvironment);
     const { user } = useUser();
     const { data: metaData } = useMeta(!!user);
